runtime enforce @final classes and faster type check
Summary:
We are starting to rely on `final` annotations on classes (e.g. we emit
`INVOKE_FUNCTION` instead of `INVOKE_METHOD` if the receiver is
final.) Although we enforce in the static compiler that final classes can't
be subclassed, this doesn't prevent non-static code from subclassing
them in violation of the annotation.

This diff marks `final` static classes as not `_Py_TPFLAGS_BASETYPE`,
so that any attempt to subclass them will be an error at runtime.

Conveniently this will give us more efficient static type checks for free,
since our JIT codegen for static type checks already does not emit the
MRO loop if `_Py_TPFLAGS_BASETYPE` is not set.

Reviewed By: czardoz

Differential Revision: D28752033

fbshipit-source-id: ab08233d79