don't reuse tmp names for inlined args
Summary:
Reusing tmp names for multiple inlined calls within the same
function is problematic because it can cause the second call to behave
differently than the first due to having an already-declared type for
the tmp name. The safest option is just to ensure that we always
increment a counter for every inline call and use that counter in the
tmp name.

Reviewed By: czardoz

Differential Revision: D28764717

fbshipit-source-id: d132bbda4c