Don't allow primitives to be used in contexts where objects are expected
Summary:
This is an issue which came up when converting a module to static, we were allowing primitive values in a context where they weren't allowed, specifically a list literal when loading from a primitive field.  The issue wasn't with the list literal code, but rather the field loading code not checking that it's type context was dynamic. There were several other places where this could happen too - calling `clen` or loading a primitive from a `Vector`.

To ensure these don't happen we really need to centralize these type checks, so I've made `set_type` now take the type context, to force flowing in the context wherever it's available.  There's still some spots where we don't have a `type_ctx`, in which case we explicitly pass `None`.

This actually gets rid of a few spots where we were doing explicit calls to verify the types of things against the `type_ctx`, now it's all handled when we do the `set_type`.

Reviewed By: carljm

Differential Revision: D28698373

fbshipit-source-id: 9b195619ac