avoid confusing BlockCanonicalizer with extra temps
Summary:
This was a fun multi-day bug hunt.

We have this `BlockCanonicalizer` in the HIR builder which ensures that
at basic block transitions the stack always consists of the first N
(where N is current stack size) allocated temporary registers.

But recently we added another `temps_.Allocate()` in early translation
setup, to allocate a null register for use as the contents of
not-yet-initialized cells, and as the replacement in `locals` array when
the locals entry is unused because it is also a cell.

This somewhat confused the block canonicalizer, because it would now
always grab this "null register" as the one that should be used for the
first item on the stack, anytime we had a non-empty stack on a block
transition.

This turned out to be mostly harmless in practice, which allowed it to
sneak through with no failures. It does mean that on deopt we would
populate this "unused" locals slot in the frame with some random object
that had been on the stack at the most recent block transition. This
isn't even directly visible to managed code, since if you ask for
`f_locals` dict, the unused `locals` slot will be ignored in favor of
the contents of the corresponding cell.

But the assignment of this random stack object to the unused `locals`
slot _does_ have the side effect of giving that random stack object
another reference, and thus keeping it alive as long as the frame lives.
So the effect becomes visible if the random stack object happens to be
something whose deallocation timing matters, such as a generator with
finalization (i.e. a generator that internally uses a context manager or
`finally` block.)

This ended up failing the exact same function in IG server that had
originally prompted us to add stack-clearing on unwind in the first
place.

This implicit and very non-obvious relationship between the temp
register allocator and the reuse of registers for stack values seems
like a bug magnet; it's already the reason for the existence of
`allocateNextTruthy`. The ideal fix is to lower directly to SSA. For now
I've just done the minimal bugfix, which is to rename
`allocateNextTruthy` to the more generic `AllocateNonStack` and add one
more call site. If we aren't going to lower directly to SSA soon, there
should probably be some better intermediate fix to make
`temps_.Allocate()` itself less bug-attractive.

Reviewed By: swtaarrs

Differential Revision: D28774788

fbshipit-source-id: bbf8036f98