Add jit support for BUILD_CHECKED_MAP
Summary:
We're missing support for the `BUILD_CHECKED_MAP` opcode in the JIT.  This adds the support for it, which is pretty much identical to the normal MakeDict except we have a specific type for the checked dict.

To be able to use the normal `SetDictItem` we need to go to `_PyDict_SetItem` instead of `_PyDict_SetItem` which does the type check.  Instead of fixing our helper call to call the right function this also just eliminates the helper call, as we've had support for returning / checking int32's for a while.

Reviewed By: carljm

Differential Revision: D28843756

fbshipit-source-id: 85d975ccba