Cleanup more JIT invoke helpers
Summary:
Given that I was touching one of these in the previous diff, I thought I'd cleanup the low-hanging fruit.  This gets rid of the invoke helpers used for int -> None/null conversion, and directly calls the function which returns negative.

It doesn't get rid of all of them though, because some of the helpers have taken on more complicated semantics than just the return value translation.

Reviewed By: carljm

Differential Revision: D28845581

fbshipit-source-id: 7170045fa7