Avoid removing call LIR instructions that may have side effects
Summary:
The register allocator removes LIR instructions with unused outputs. However, some instructions may have side effects and should be kept. For example, a function call to `PyErr_Format` returns `NULL` which is likely to be unused. However, we expect the function to print an error message.

This change in `Jit/codegen/regalloc.cpp` will keep any `kCall` or `kVectorCall` instruction regardless if the output is used or not. There is also a new test `CallWithSideEffectTest` in `LinearScanAllocatorTest` to check that calls are being kept.

Reviewed By: swtaarrs

Differential Revision: D28850668

fbshipit-source-id: 87ee61b12a