Fix unboxing int with conversion in place
Summary:
There's a little more fall out from the additional type checking in `set_type`.  We now don't allow the pattern of `int64(int(some_value))` because we end up with an `INT_EXACT_TYPE` for the conversion to `int64`, but it doesn't have a literal value.  So we fail the check in `is_valid_literal_int`.

But we also can't just can't let all `INT_EXACT_TYPE` go through here - that'll break error reporting on something like `int8(255)`.  So now we need to start tracking the literal types for integers to be able to make this work.

This only adds literal tracking of int's...  to support more types we'll probably want to update how we convert an exact type back into an inexact type.  Right now it just becomes a special case :(

Reviewed By: carljm

Differential Revision: D28937606

fbshipit-source-id: 875e760300