support for unboxing doubles
Summary:
This adds support for unboxing floating point values, in the JIT and
interpreter.

One thing I needed to change was the `IsNegativeAndErrOccurred ` HIR opcode. Currently, it only operates on an integer argument, but it had to be updated to support doubles.

I guess we could also entirely remove the call to `JITRT_UnboxDouble`, and push `PyFloatObject->ob_fval` on the stack, but as always, we might just get a float subclass here.

Reviewed By: DinoV

Differential Revision: D28840938

fbshipit-source-id: c5046999bb