Attempt to resolve type and make types non-GC
Summary: When we setup a type with `__slot_types__` we have the possibility of creating a no-GC type.  But for any field declared as an object we currently restore to GC.  This causes us to attempt to resolve the type (it could fail because its a forward reference), and if we can, then check to see if its a GC type and if it's final.  If so we don't need to make the existing type GC tracked.

Reviewed By: carljm

Differential Revision: D28887236

fbshipit-source-id: d3776993bc