Apply narrowing effects for `assert test, message`
Summary:
Does what it says on the tin.

I had a question about the effect system: Are statements expected to propagate the effects that they've applied, or is the action of applying the effect on locals enough for our purposes here? The terminology is a bit different from what I'm used to is why I'm asking.

Reviewed By: carljm

Differential Revision: D29032886

fbshipit-source-id: f6a5c6fa57