Preserve order of basic block successors in LIR printer
Summary:
The basic block successors should be printed in the order that they appear in the successors vector, instead of being printed in sorted order. This is because the order of the successors determines which block to jump to during conditional branching. So, when we are parsing the printed functions, we need the order of successors to correspond to the branching order (i.e. the first successor is the true branch, the second successors is the false branch).

Note that we will still sort the predecessors before printing to ensure consistency.

We also update the LIR translations in `Jit/lir/c_helper_translations/` with the correct order of successors.

Differential Revision: D29213165

fbshipit-source-id: 9f59777ede