Add Simplify pass
Summary:
This is a catch-all pass for small strength-reduction optimizations,
to avoid duplicating the overhead (both lines of code and runtime cycles) of
all the boring stuff surrounding the optimization itself. I've seeded it with
all existing optimizations that fit; see the comments in `optimization.h` and
`simplify.cpp` for more details.

Reviewed By: tiansihu

Differential Revision: D29241052

fbshipit-source-id: 6d2491372c