Properly refine or's
Summary: When analyzing `e1 or e2`, we should only consider `e1`'s type in the case where the narrowing effects of `e1` hold. More concretely, when analyzing `x is not None or e2`, the result will only evaluate to `x is not None` if the condition holds, so we need to ensure that we're considering that effect when widening the type. The only expression that gets returned unconditionally is the last expression in an or, so we shouldn't apply any effects for those, but still need to build the effect as it gets consumed.

Reviewed By: DinoV

Differential Revision: D29272909

fbshipit-source-id: b11f221bf5