Teach RefcountInsertion about immortal objects
Summary:
`None`, `True`, and `False` are all immortal now so we don't need to
refcount them. Small ints and static types are as well, so we inspect object
specializations to detect those.

The actual change is just a few lines in `refcount_insertion.cpp`, then that
led to a bunch of changes to tests.

Reviewed By: carljm

Differential Revision: D29304967

fbshipit-source-id: 3bf9446579