Optimize Compare<is> to one primitive compare
Summary:
Existing optimizations rewrote the `if x is y: ...` case to
remove the `IsTruthy`. This diff rewrites the more general case where
`x is y` is used as an expression, such as in `return x is y`. This does
not have an `IsTruthy` and can be expanded into
`PrimitiveCompare`/`PrimitiveBox`.

Reviewed By: carljm

Differential Revision: D29335715

fbshipit-source-id: 3b84b11c82