Return self for exact/inexact of isinstance
Summary: For `isinstance` and other built-in functions we want the exact/in-exact of themselves to be themselves.   Right now the default behavior is to fallback to the type, which turns these into a less specific `FunctionType`.  This only manifests as a problem where the strict module re-writier assigns these values which forces us to go through `inexact`.

Reviewed By: czardoz

Differential Revision: D29404443

fbshipit-source-id: c77eec7b4e