fix chained comparisons
Summary:
Fixes several bugs with our chained comparison implementation,
when used with PyObjects and not primitives.  We were overriding
`emitChainedCompareStep` inconsistently with the base method on
`CodeGenerator`, causing errors in compilation when it was called from
the base implementation of `compileJumpIf`. Fixed this by just
re-getting `optype` instead of passing it in.

With that fixed, we would give wrong results for chained compares of
PyObjects because our `emitChainedCompareStep` always defaulted to
`JUMP_IF_ZERO_OR_POP` instead of `JUMP_IF_FALSE_OR_POP`, even though the
former is only correct for primitives; with objects, the pointer value
is always nonzero, so we would never jump. Fixed this by switching from
an explicit opcode as argument to an `always_pop` boolean, which lets
the static compiler then polymorphically call one of
`emit_jumpif_pop_only` or `emit_jumpif_only` on the optype. These
methods are extracted from `emit_jumpif_pop` and `emit_jumpif` to only
emit the jumpif opcode and not visit a test node first, since this is
how they need to be used in chained compares. This allows us to
centralize the override to use the `[NON]ZERO` opcode variants instead
of the `{TRUE,FALSE}` variants in just one place.

Ideally we would get rid of the `[NON]ZERO` variants and just let the
JIT use type information to intelligently elide the `IsTruthy` step when
not needed. The trick here is that we don't have type information at
lowering, so this would need to be done as a later pass, which would not
be an optimization but would be required for correctness. Unless we just
expand the semantics of `IsTruthy` to be that at lowering to HIR, it
becomes a no-op if its input is a primitive.

Reviewed By: DinoV

Differential Revision: D29449067

fbshipit-source-id: a4bc0c1a36