Allocate Runtime
Summary:
`s_jit_asm_code_rt_` is getting cleaned up at process exit.  This really should be explicitly cleaned up as part of runtime finalization, and we have `clearReferences` which does that.

But we can't delete the runtime object when in the existing finalization callback - we'll get more callbacks as the runtime continues to finalize it's own objects.  So we add a new `_PyJIT_Finalized` which can release the resources once the runtime is fully shut down.

Reviewed By: jbower-fb

Differential Revision: D28587786

fbshipit-source-id: ef3e47ccde