Replace physical registers with load arg in C helper LIR translations
Summary:
In the LIR translation of C helper functions, we need to get the arguments. Instead of specifying physical registers, we can use the `kLoadArg` instruction. This makes it easier to identify which instructions correspond to retrieving the arguments.

Set up `environ.arg_locations` in `SimpleCompile` to allow rewriting of `kLoadArg` instructions in post gen.

Updated the load instructions used in `JITRT_Get32_FromArray` and `JIRT_Cast` translations in `RuntimeTests/backend_test.cpp` and `Jit/lir/c_helper_translations/`.

Reviewed By: swtaarrs

Differential Revision: D29401662

fbshipit-source-id: 07cb9d98c6