Revert "Fix refcounting bug in gen_send variants"
Summary:
This reverts commit d027502c0f9325526445f89dc2b62e982fffcfc6, or D29461030 (https://github.com/facebookincubator/cinder/commit/dd220fe0073ed43417e755076a449ce332a92d74).

That diff reliably causes segfaults in various places in shadowcode and GC collection, when
deployed to prod or run on a devserver. I haven't identified the precise bug but I presume
it's a use-after-free due to something being over-decrefed and collected too soon.

Blame diff identified by bisect, and confirmed that latest `cinder/3.8` with this revert does
not have the segfaults.

Reviewed By: swtaarrs

Differential Revision: D29609615

fbshipit-source-id: 1634fd3e12