Move primitive_unbox + load optimization into jit
Summary:
The goal of this diff is to generalize the constant unbox optimization so that we no longer need to special case `NumClass` in the static compiler.

The specific optimization eliminates box/unboxing of a value with a known Object spec and rewrites it as a `LOAD_CONST`. This is useful for JustKnobs and Gatekeepers, as they load `RAND_MAX` in this fashion a lot (this was the motivation for the special case).

The diff is currently WIP, publishing early to (a) get feedback on direction (b) so that I can concretely have a few questions I have.

Reviewed By: carljm

Differential Revision: D29106120

fbshipit-source-id: affa0f88cf