Introduce an HIR optimization pass to eliminate redundant error checks
Summary:
Other optimizations in static python, such as const unbox elimination, can make negative + error occurred checks redundant. This pass replaces the check with a const load that is hopefully trivial to optimize out/inline based on need in the LIR.

In order to test this, I had to add support for multiple optimization passes running in a single HIR test as it doesn't make sense to test these in isolation, so that scaffolding was added.

This change is part of the work needed to remove the static python unbox optimization special case.

Reviewed By: carljm

Differential Revision: D29136880

fbshipit-source-id: 486748f61f