modify codegen of static type checks to omit checks for defaulted arguments
Summary:
Context: The way we handle defaults right now is via JITRT_CallWithIncorrectArgcount where we fill the argspace for missing args from the callee's defaults. For non-static -> static calls, this means that we will re-emit the type checks for each argument unnecessarily. This diff attempts to reduce this overhead by modifying the codegen of the callee to add a section which dispatches to the appropriate type check depending on the number of locals.

The basic strategy is that we now have a label preceding every check that's generated beforehand, and each local up to the first appropriate check (which can be multiple due to eliding checks for dynamically typed locals) have a `cmp number_of_defaulted; jump_to_appropriate_check` pair of instructions that'll skip all the checks before. To allow this strategy to work, we also reorder the checks in JIT'ed code to start with the last argument, and check the first one last.

RFC:

1. I'm building a python package and setting up swayze tests now - is there any other perf tests you'd recommend?
2. Do people have opinions on how to name/clean up the index math a bit? I want to keep things as readable as possible for people, and would appreciate opinions as I think it's complex at the moment.

Reviewed By: DinoV

Differential Revision: D29525681

fbshipit-source-id: edeb0b80af