Adds error sink that can gather errors
Summary:
This adds a new `ErrorSink` API which can gather the errors from the static compiler as they're processed.  For the most part this is a pretty straight forward move the old error reporting to call the error sink.  There's a few changes which fall out of it:
1) We now need to handle the case where the error sink doesn't raise in a number of places.  There's various strategies such devolving to the dynamic type, returning a bool to indicate operations failed, returning `NO_EFFECT` from potentially narrowing operations, etc...
2) There's some cleanup to stop reporting errors from the code generator and instead push those errors to the binder.  This requires us to no longer set the dynamic type on a failed forward comparison, have a special helper for unions for comparison operations to avoid getting a type when a compare wasn't processed, allow dynamic to be passed to `double`, and moving the primitive assignment from double check into `check_can_assign_from`

It's also becoming apparent that SymbolTable has kind of grown beyond being a symbol table :P  It's had a `compile` method on it for a while, not it gets the `ErrorSink` flowed into it because we don't have any other good places, and it gets the `bind` method which only does the type binder.  Because of this I've added an alias for it as `Compiler` so we can tease out the two pieces in the future.

Finally the checks on the new tests are extremely loose - that's because we're totally over-reporting errors right now.  On an assignment like `x: int = 'abc'` we'll report an error when we visit the string in an int context, and then we'll report another error when we attempt the assignment.  While it gives some nice assurances that we're doing the type checking with the type context, it seems like it'll make our errors a mess.  My initial pass to fix this also started having better error messages (e.g. something like "incorrect return type X, expected Y" instead of "you can't assign X to Y").  So I think we need to revisit that design decision and always check types on assignment and truly just use the type context as a hint.  Then we can tighten up the checks for the expected error messages.

There's also unfortunately a lint error reported where it wants to remove our imports of `ErrorSink` and `TypedSyntaxError`.  The latter is simply exported for convenience and compatibility, but `ErrorSink` is actually used!

Reviewed By: czardoz

Differential Revision: D29522619

fbshipit-source-id: 69f2e9b520