Cleanup check if a the paramter to CInt is valid
Summary:
The logic here is sound, but it's grown rather organically, and reads really poorly.  This breaks it out into several `if`s where it's much more obvious what's going on.

The next couple of diffs were touching this and moving it around so this just pre-emptively cleans it up.

Reviewed By: czardoz

Differential Revision: D29624941

fbshipit-source-id: 626090bcc5