Don't enforce types in set_type anymore
Summary:
This is basically backing away from the approach used in D28698373 (https://github.com/facebookincubator/cinder/commit/695adea62c085209e3a4bef7fc66a41e2f7e20a2).  It turns out `set_type` isn't a very good place to centralize these checks.  There's a few of problems:
1) With an error sink used for error reporting, we end up reporting multiple errors when we are otherwise doing type checks later.  First we report an error about the `type_ctx` in `set_type`, and then we report an error when we check the error.
2) Related to #1 we also have a pattern that's developed of `try/except TypedSyntaxError:` to improve the error messages.  This also leads to duplicate messages about the same error.
3) Without a `type_ctx` we end up reporting a worse error than we could if we check later.  You can see that in this diff where a lot of errors in test cases go from a random type mismatch to an unexpected return type.
4) We typically flow down the existing `type_ctx` and don't introduce a new one.  We could start doing `visit(..., type_ctx or DYNAMIC)` but given the first two issues that doesn't seem like that's the right solution.

So this diff introduces a new `visitExpectedType` which visits the node, and then validates that the type is assignable with the option for a custom error message.  This API doesn't necessarily have to be used consistently because we frequently do type checks for other calls.

It removes the type check in `set_type` but doesn't yet update the callers to remove passing the parameter, that'll come in a separate diff to keep things simple.  That'll be done in the next diff.

Reviewed By: carljm

Differential Revision: D29577188

fbshipit-source-id: 84da68781e