Literal int primitive constants
Summary:
This is continuing to cleanup the pattern of `try/except TypedSyntaxError:`.  Specifically it's adding suppor for literal CIntType's so that we no longer do the enforcement of constant ranges in `bind_constant`.  Instead we construct a literal CInt and we make sure that we have a valid range when we attempt to assign it.

One interesting question might be what should we display these as - I've kept it the same as a normal `Literal[###]` like we do for normal ints.  And I think looking at the changes in the error messages this looks okay, but am open to feedback.

After this diff the only place where we attempt to recover from a syntax error and report another one is for union instances.  Arguably this isn't the worse behavior - the user will get a report that the individual elements are wrong, and then that the union type is wrong.  Currently we'd only report on the first wrong type though.

Reviewed By: sinancepel

Differential Revision: D29577187

fbshipit-source-id: 96f1213564