Add missing visitExpectedType calls and tests
Summary:
This goes through our remaining `visit` calls and replaces them with `visitExpectedType` calls where we specifically pass in `DYNAMIC` and adds test cases that we're not allowing primitives where we shouldn't be.  Most of these represents holes in our existing analysis that result in crashes.

One thought I had half way through this was if we should be passing `DYNAMIC` here or `OBJECT`.  It feels like in general we should be using `OBJECT` as what we really want is something that is in fact an object.  For consistency sake I left this as `DYNAMIC` but could follow up and switch everything to `OBJECT` if others agree that sounds better.

Reviewed By: carljm

Differential Revision: D29603200

fbshipit-source-id: 70413da04d