Add a method that splits a block before a given instruction
Summary:
Added a method `splitBefore` that splits a block before a given instruction. The original block should contain all the instructions up to but excluding the given instruction. The method returns a new block that contains all of the instructions including and after the given instruction.

We also fix up phi instructions accordingly. After splitting the block, the function (that the block belongs to) should still behave the same.

Differential Revision: D29551315

fbshipit-source-id: 6700d2a49e