move NarrowingEffects into their own file
Summary:
This is part of the goal of shaving off bits of the static compiler into logical, self-contained files.

The remaining work here is to also move IsInstanceEffect here. There's an ugly mutual recursion with `UNION_TYPE` <-> `IsInstanceEffect` that will probably be resolved once we can come up with a way of moving all type definitions into their own module.

Reviewed By: DinoV

Differential Revision: D29739289

fbshipit-source-id: b7546c112e