initial support for checked lists
Summary: This is a scaffolding diff that aims to set up a basic `chklist` that we can build on in the future. For now, simply supports creation with a single type parameter, and acts as a normal list afterwards. Future diffs will add type checks on insertion from non-static code, and start eliding checks in the static cases.

Reviewed By: czardoz

Differential Revision: D29831377

fbshipit-source-id: 32f7f8478b