static method consistency
Summary: Ensures that enforce override consistency of static methods as well.  This also currently disables static dispatch via an instance method to a static method, and this is wrong in the face of overrides.  Theoretically a derived class could override the static method, in which case we should invoke via the instance.

Reviewed By: sinancepel

Differential Revision: D29726914

fbshipit-source-id: e12c1b3ea3