Add support for dynamically resolving static methods when invoked via instances
Summary:
The previous change disabled our ability to directly invoke static methods statically when invoked via an instance.  That's because the target of the invoke could change - the static method could be overridden in a subclass.

This adds support for static methods in vtables so that we can generate an `INVOKE_METHOD` against them, and then we drop the instance upon calling the final function.

Reviewed By: czardoz

Differential Revision: D29804656

fbshipit-source-id: 6f34bb44bf