Support primitive default values
Summary:
This was just something I noticed while I was cleaning up the type enforcement for primitive values.  We don't do the correct visit with a type for our default values.  There's some additional downstream fallout in that when emitting the default value for a call where it's not supplied we also don't get the correct type there, and instead do a generic `LOAD_CONST`.

So now we visit with the correct type at the definition, and then we create a new constant for the value to be emitted and visit with the correct type there too.  We need to make sure that we emit the value for `__defaults__` as a boxed value though still, so there's a new visitDefault added to the base code generator.

This exposed one issue in the JIT which was dependent upon how things were compiled.  When we have a primitive return value, and primitive args, the combination wasn't necessarily working well.  In that case we have a small call to the primitive returning function so we can do an unbox after, and check edx for error.  After making the small call into the JITed code for the return we call `JITRT_CallStaticallyWithPrimitiveSignature` to unbox the primitive argument types.  But this function is typed to return a PyObject*.  So we're now smuggling the primitive return value through a PyObject*, and we can trash rdx on the way back (likely from various `Ref<>` holders we have).

So now `JITRT_CallStaticallyWithPrimitiveSignature` is updated to use `JITRT_StaticCallReturn` as well, and we can return either return a PyObject* in rax and rdx isn't important, or we can return a primitive in rax and rdx properly has the error code.

Reviewed By: czardoz

Differential Revision: D29603798

fbshipit-source-id: a6cce2fdbb