Create a deep copy of a function's blocks into another function
Summary:
Added a method `copyInto` that creates a deep copy of the current function's blocks into `dest_func` in between the blocks `prev_bb` and `next_bb`. Note that the id's of the new blocks and instructions are automatically assigned.

Added constructors for `Instruction` and `Operand` classes.

Added runtime tests `CopyFromArrayTest` and `CopyCastTest`. Cleaned up the backend tests so that we can call helper functions `CheckFromArray` and `CheckCast` to test the behavior of `JITRT_GetI32_FromArray` and `JITRT_Cast`.

Reviewed By: swtaarrs

Differential Revision: D29572182

fbshipit-source-id: fbd522531d