Add findFunction method to LIRInliner class
Summary:
Added `findFunction` method which takes a call instruction and attempts to find the corresponding callee function. This function returns a nullptr if the callee cannot be found.

We use dladdr to resolve the address from the call instruction to a function name. Then, we attempt to find the hand-translated lir that corresponds to the function name and parse the text.

Added `FindFunctionSuccessTest` and `FindFunctionFailureTest` to test `findFunction`.

Differential Revision: D29741833

fbshipit-source-id: 1ed5832b13