Add InlineCall and isInlineable methods to LIRInliner class
Summary:
Added `InlineCall` and `isInlineable` methods to LIRInliner class.

`InlineCall` is the public method that we call in order to inline a call instruction (`call_instr_`) . It checks that the callee is inlineable and performs argument/return value resolution. If inlining succeeds, `InlineCall` will return true and the `call_instr_` will be replaced with the callee. We can call this method from a post gen pass to inline helper functions.

`isInlineable` is a private method that we use to check if the `call_instr_` and the `callee` is inlineable. It calls `checkEntryExitReturn` which checks that there is exactly 1 entry and 1 exit block and that return statements only appear right before an exit block. It calls `checkArguments` which checks that arguments in `call_instr_` are immediates or virtual registers. It calls `checkLoadArg` which checks that the load arg instructions in `callee` are valid.

We removed the separate pass to track the arguments in `resolveArguments`. Instead, `checkArguments` will add the elements into a member variable `arguments_` that we can use in `resolveArguments`.

Reviewed By: tiansihu

Differential Revision: D29803610

fbshipit-source-id: 0dbf6a3353