Add stub support
Summary:
Add support for stubs and the `implicit` decorator.
The module loader has a separate path in which it will look for stubs (`pys` files).
How `implicit` works:
- If no implicit exist in a stub, the stub need not have a backing source file.
- The source file backing a stub will be parsed into C AST when the stub containing `implicit` is analyzed. A mapping from function names to AST nodes in the source will be created. If there is ambiguity, an error is raised.
- Another mapping from AST locations (indices into asdl_seq) to names is built from the stub.
- The AST location will be replaced with the corresponding AST.

In order to support test cases, a subset of stubs is ported into StrictModules/Tests. Other than stubs necessary for testing, the rest of the stubs can live anywhere and the path is set by user.

Reviewed By: carljm

Differential Revision: D28395668

fbshipit-source-id: 93d50654ce