Add dataclass stub
Summary:
Add a stub for dataclasses. This replaces the hand written dataclass implementation in the python strict module analyzer.

To support usage of deepcopy, all strict objects now support a `copy` method.
Added builtin support to inspect sys.modules and get the value if it exists. This is implemented as a separate builtin that's called in stub.

All places in the stub where `implicit` are not used are commented with a reason for modifying the source code. Most notably, checks and implementations specific to frozen dataclasses are omitted (i.e. for analysis purpose we assume dataclasses are not frozen) because we don't support `__getattribute__`, and that the behavior is the same for analysis purposes.

Reviewed By: carljm

Differential Revision: D28574714

fbshipit-source-id: c954926aba