Handle int overflow
Summary:
Python handles arbitrarily long integers but in hour modeling of python ints we used C++ longs, which is only 64 bit.
To avoid overflow, StrictInt now optionally contains a C++ representation. When int is too large, the PyObject will be used instead.
Note that this does not fully solve the problem: two StrictInt with C++ representations can result in overflow, which will not be caught. I don't think this is a concern for our analysis because 1. big ints are rare and mostly not used in arithmetics in our workload, and 2. concrete value of ints rarely matters for the analysis. But this may need revisiting.

Reviewed By: carljm

Differential Revision: D28841257

fbshipit-source-id: 92719a9edd